pyproject.toml
[project]
name = "weather_app"
version = "0.1.0"
description = "Weather app collect and correct field measurements along with a mock server"
requires-python = ">=3.10"
dependencies = [
    "openapi-python-client == 0.26.1",
    "pydantic",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pytest"
]

[package]
name = "weather_app"
version = "0.1.0"
source = { path = "src/weather_app", relative = true }

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests/"]
python_files = ["test.py"]

